requirements.txt
Django<4
djangorestframework
djangorestframework-jwt
psycopg2-binary
